Primera Version con registro de student para con funcionalidad para de una vez iniciar sesion

+++ app/src/main/java/oportunia/maps/frontend/taskapp/presentation/navigation/NavGraph.kt
@@ -393,7 +393,7 @@ fun NavGraph(
     studentViewModel: StudentViewModel,
     qualificationViewModel: QualificationViewModel,
     paddingValues: PaddingValues,
-    onRegisterSuccess: () -> Unit
+    onRegisterSuccess: (Int) -> Unit
 ) {
     NavHost(navController = navController, startDestination = NavRoutes.MainRegister.ROUTE) {
